phase3_llamaindex/requirements.txt
# ============================================
# Phase 3: LlamaIndex - 依赖包
# ============================================
# 安装: pip install -r requirements.txt

# ============================================
# LlamaIndex 核心
# ============================================
llama-index==0.9.27
llama-index-core==0.9.27

# ============================================
# LLM 集成
# ============================================
llama-index-llms-openai==0.1.8
llama-index-llms-anthropic==0.1.11

# ============================================
# Embeddings 集成
# ============================================
llama-index-embeddings-openai==0.1.5
llama-index-embeddings-huggingface==0.1.1

# ============================================
# 向量数据库
# ============================================
llama-index-vector-stores-chroma==0.1.5
llama-index-vector-stores-faiss==0.1.10
chromadb==0.4.22
faiss-cpu==1.7.4

# ============================================
# 数据加载器
# ============================================
llama-index-readers-file==0.1.15
llama-index-readers-web==0.1.11
llama-index-readers-database==0.1.6

# ============================================
# 文档处理
# ============================================
pypdf==3.17.4
python-docx==1.1.0
beautifulsoup4==4.12.3
html2text==2020.1.16

# ============================================
# 其他依赖
# ============================================
python-dotenv==1.0.0
tiktoken==0.5.2
